move view allocation related functionality to a new header
